Fix deaths registering twice when inside of a car
This is a fix for deaths registering twice whilst dying inside of a car by someone else. If someone killed you with a sword and you were riding in a car it will show 2 deaths rather then one. Not sure why, but I assume it has to do with using damageable and damaging the player. This might not be an issue on 1.7, but tested on 1.8.3 as an issue.